move model and layer_spec to unit tests
