+++ src/paddlefleet/models/qwen3_5/__init__.py
@@ -13,27 +13,19 @@
 # limitations under the License.
 
 from .layer_specs import (
-    get_qwen3_5_language_spec,
     get_qwen3_5_vision_spec,
 )
 from .qwen3_5_model import (
-    FleetQwen3_5ForConditionalGeneration,
-    Qwen3_5Model,
+    Qwen3_5RMSNorm,
+    Qwen3_5RMSNormPipe,
     Qwen3_5VisionModel,
     Qwen3_5VisionSublayersSpec,
 )
-from .qwen3_5_provider import (
-    Qwen3_5VisionProvider,
-    Qwen3_5VLProvider,
-)
 
 __all__ = [
-    "get_qwen3_5_language_spec",
     "get_qwen3_5_vision_spec",
-    "Qwen3_5Model",
+    "Qwen3_5RMSNorm",
+    "Qwen3_5RMSNormPipe",
     "Qwen3_5VisionModel",
     "Qwen3_5VisionSublayersSpec",
-    "Qwen3_5VisionProvider",
-    "Qwen3_5VLProvider",
-    "FleetQwen3_5ForConditionalGeneration",
 ]
