+++ tests/single_card_tests/model/test_qwen3_5_vision_model.py
@@ -13,18 +13,562 @@
 # limitations under the License.
 
 
+import itertools
 import random
 import unittest
+from dataclasses import dataclass
 
 import numpy as np
 import paddle
+from paddle import Tensor
 from paddle.distributed import fleet
+from paddle.nn import functional as F
 
 import paddlefleet.parallel_state as ps
+from paddlefleet.models.common.empty_layer import EmptyLayer
 from paddlefleet.models.gpt import GPTConfig
-from paddlefleet.models.qwen3_5.qwen3_5_model import Qwen3_5Model
-from paddlefleet.models.qwen3_5.qwen3_5_provider import Qwen3_5VisionProvider
+from paddlefleet.models.gpt.gpt_embedding import GPTEmbedding
+from paddlefleet.models.gpt.gpt_layer_specs import (
+    get_gpt_layer_local_spec,
+    get_gpt_spec,
+)
+from paddlefleet.models.gpt.lm_head import GPTLMHead
+from paddlefleet.models.qwen3_5.layer_specs import get_qwen3_5_vision_spec
 from paddlefleet.pipeline_parallel import NoPipelineParallel
+from paddlefleet.spec_utils import LayerSpec, build_layer
+from paddlefleet.tensor_parallel.mappings import (
+    scatter_to_sequence_parallel_region,
+)
+from paddlefleet.transformer import TransformerConfig
+from paddlefleet.transformer.layer import FleetLayer
+from paddlefleet.transformer.paddle_norm import (
+    WrappedPaddleNorm,
+    WrappedPaddleNormPipe,
+)
+from paddlefleet.utils import get_tensor_model_parallel_group_if_none
+
+# ======================================================================
+# Qwen3_5VisionProvider (inlined from deleted qwen3_5_provider.py)
+# ======================================================================
+
+
+@dataclass
+class Qwen3_5VisionProvider(TransformerConfig):
+    patch_size: int = 16
+    use_bias: bool = True
+    add_qkv_bias: bool = True
+    num_position_embeddings: int = 2304
+    embed_dim: int = (1152,)
+    hidden_size: int = 1152
+    out_hidden_size: int = 3584
+    in_channels: int = 3
+    spatial_merge_size: int = 2
+    spatial_patch_size: int = 16
+    temporal_patch_size: int = 2
+    hidden_dropout_prob: float = 0.0
+    attention_dropout: float = 0.0
+    intermediate_size: int = 4304
+    initializer_range: float = 0.02
+    gated_linear_unit: bool = False
+    activation_func: object = F.gelu
+    layernorm_zero_centered_gamma: bool = False
+    apply_query_key_layer_scaling: bool = False
+    persist_layer_norm: bool = True
+    bias_activation_fusion: bool = False
+    bias_dropout_fusion: bool = False
+    attention_softmax_in_fp32: bool = True
+    normalization: str = "LayerNorm"
+    apply_rope_fusion: bool = True
+    rms_norm_eps: float = 1e-6
+    model_version: str = "qwen3_5"
+
+    def provide(self):
+        spec = get_qwen3_5_vision_spec(self)
+        return build_layer(
+            spec,
+            seg_method="layer:TransformerLayer|EmptyLayer",
+            num_stages=self.pipeline_model_parallel_size,
+        )
+
+
+# ======================================================================
+# get_qwen3_5_language_spec (inlined from deleted layer_specs function)
+# ======================================================================
+
+
+from paddlefleet.models.qwen3_5.qwen3_5_model import (
+    Qwen3_5RMSNorm,
+    Qwen3_5RMSNormPipe,
+)
+
+
+def get_qwen3_5_language_spec(config):
+    layer_types = getattr(config, "layer_types", None)
+    if layer_types is None:
+        layer_types = ["full_attention"] * config.num_hidden_layers
+
+    empty_layer_spec = LayerSpec(
+        layer=EmptyLayer, extra_kwargs={"config": config}
+    )
+    head_empty_layers = [empty_layer_spec] * config.num_empty_layers_add_in_head
+    tail_empty_layers = [empty_layer_spec] * config.num_empty_layers_add_in_tail
+
+    head_offset = getattr(config, "num_empty_layers_add_in_head", 0)
+
+    LAYER_TYPE_MAP = {
+        "full_attention": "self_attention",
+        "linear_attention": "gated_delta_net",
+    }
+
+    transformer_layers_spec = []
+    for i, lt in enumerate(layer_types):
+        attn_type = LAYER_TYPE_MAP.get(lt)
+        if attn_type is None:
+            raise ValueError(f"Unknown layer type: {lt!r} at index {i}")
+        spec = get_gpt_layer_local_spec(
+            config=config,
+            normalization=config.normalization,
+            layer_number=i + head_offset,
+            attention_layer_type=attn_type,
+            num_experts=config.n_routed_experts,
+            moe_grouped_gemm=config.moe_grouped_gemm,
+        )
+
+        sub = spec.sublayers_spec
+        if sub.input_layernorm is WrappedPaddleNorm:
+            sub.input_layernorm = Qwen3_5RMSNorm
+        if sub.post_attention_layernorm is WrappedPaddleNorm:
+            sub.post_attention_layernorm = Qwen3_5RMSNorm
+
+        attn_spec = sub.self_attn
+        if hasattr(attn_spec, "sublayers_spec"):
+            attn_sub = attn_spec.sublayers_spec
+            if (
+                hasattr(attn_sub, "q_norm")
+                and attn_sub.q_norm is WrappedPaddleNorm
+            ):
+                attn_sub.q_norm = Qwen3_5RMSNorm
+            if (
+                hasattr(attn_sub, "k_norm")
+                and attn_sub.k_norm is WrappedPaddleNorm
+            ):
+                attn_sub.k_norm = Qwen3_5RMSNorm
+
+        transformer_layers_spec.append(spec)
+
+    full_spec = get_gpt_spec(
+        config=config,
+        transformer_layers_spec=transformer_layers_spec,
+        mtp_layers_spec=None,
+        vocab_size=config.vocab_size,
+        max_sequence_length=config.max_sequence_length,
+        head_empty_layers_spec=head_empty_layers,
+        tail_empty_layers_spec=tail_empty_layers,
+        position_embedding_type=config.position_embedding_type,
+        rotary_percent=config.rotary_percent,
+        rotary_base=config.rotary_base,
+        rope_scaling=config.rope_scaling,
+        parallel_output=config.parallel_output,
+        tie_word_embeddings=config.tie_word_embeddings,
+    )
+
+    final_norm_spec = full_spec.sublayers_spec.layer_norm
+    if final_norm_spec.layer is WrappedPaddleNormPipe:
+        final_norm_spec.layer = Qwen3_5RMSNormPipe
+
+    return full_spec
+
+
+# ======================================================================
+# Qwen3_5Model (inlined from deleted qwen3_5_model.py class)
+# ======================================================================
+
+
+class Qwen3_5Model(FleetLayer):
+    def __init__(
+        self,
+        config,
+        vision_model=None,
+        language_model=None,
+        spatial_merge_size=2,
+        image_token_id=None,
+        video_token_id=None,
+    ):
+        assert isinstance(language_model, NoPipelineParallel)
+        assert isinstance(vision_model, NoPipelineParallel)
+        super().__init__(config=config)
+        self.visual = vision_model
+        self.language_model = language_model
+        self.spatial_merge_size = spatial_merge_size
+        self.image_token_id = image_token_id
+        self.video_token_id = video_token_id
+        self.rope_deltas = None
+
+        self.language_embedding = self._find_language_embedding()
+        self.language_backbone = self._find_language_backbone()
+        self.language_lm_head = self._find_lm_head()
+
+        self.tp_group = get_tensor_model_parallel_group_if_none(None)
+
+        if self.language_embedding is not None:
+            embed_tokens = self.language_embedding.embedding.embed_tokens
+            embed_tokens.reduce_scatter_embeddings = False
+
+    def _find_language_embedding(self):
+        for layer in self.language_model._layers.run_function:
+            if isinstance(layer, GPTEmbedding):
+                return layer
+        return None
+
+    def _find_language_backbone(self):
+        return [
+            layer
+            for layer in self.language_model._layers.run_function
+            if not isinstance(layer, (GPTEmbedding, GPTLMHead))
+        ]
+
+    def _find_lm_head(self):
+        for layer in self.language_model._layers.run_function:
+            if isinstance(layer, GPTLMHead):
+                return layer
+        return None
+
+    def get_image_features(self, pixel_values, image_grid_thw=None, **kwargs):
+        dict_input = {
+            "pixel_values": pixel_values,
+            "grid_thw": image_grid_thw,
+        }
+        output = self.visual._layers.forward(dict_input)
+        if isinstance(output, tuple):
+            return output[0]
+        return output
+
+    def get_video_features(
+        self, pixel_values_videos, video_grid_thw=None, **kwargs
+    ):
+        return self.get_image_features(
+            pixel_values_videos, video_grid_thw, **kwargs
+        )
+
+    def get_placeholder_mask(
+        self,
+        input_ids,
+        inputs_embeds,
+        image_features=None,
+        video_features=None,
+    ):
+        if input_ids is None:
+            embed_fn = self.get_input_embeddings()
+            special_image_mask = (
+                inputs_embeds
+                == embed_fn(
+                    paddle.to_tensor(self.image_token_id, dtype="int64")
+                )
+            ).all(-1)
+            special_video_mask = (
+                inputs_embeds
+                == embed_fn(
+                    paddle.to_tensor(self.video_token_id, dtype="int64")
+                )
+            ).all(-1)
+        else:
+            special_image_mask = input_ids == self.image_token_id
+            special_video_mask = input_ids == self.video_token_id
+
+        n_image_tokens = special_image_mask.sum()
+        special_image_mask = special_image_mask.unsqueeze(-1).expand_as(
+            inputs_embeds
+        )
+        if image_features is not None:
+            assert int(inputs_embeds[special_image_mask].numel()) == int(
+                image_features.numel()
+            )
+
+        n_video_tokens = special_video_mask.sum()
+        special_video_mask = special_video_mask.unsqueeze(-1).expand_as(
+            inputs_embeds
+        )
+        if video_features is not None:
+            assert int(inputs_embeds[special_video_mask].numel()) == int(
+                video_features.numel()
+            )
+
+        return special_image_mask, special_video_mask
+
+    def get_vision_position_ids(
+        self,
+        start_position,
+        grid_thw,
+        spatial_merge_size=1,
+        device=None,
+    ):
+        if isinstance(grid_thw, Tensor):
+            t = int(grid_thw[0].item())
+            h = int(grid_thw[1].item())
+            w = int(grid_thw[2].item())
+        else:
+            t, h, w = int(grid_thw[0]), int(grid_thw[1]), int(grid_thw[2])
+
+        llm_t = t
+        llm_h = h // spatial_merge_size
+        llm_w = w // spatial_merge_size
+        seq_len = llm_t * llm_h * llm_w
+
+        pos_w = paddle.arange(start_position, start_position + llm_w).tile(
+            [llm_h * llm_t]
+        )
+        pos_h = paddle.arange(
+            start_position, start_position + llm_h
+        ).repeat_interleave(llm_w * llm_t)
+        pos_t = paddle.full([seq_len], start_position, dtype="int64")
+
+        return paddle.stack([pos_t, pos_h, pos_w], axis=0)
+
+    def get_rope_index(
+        self,
+        input_ids,
+        mm_token_type_ids,
+        image_grid_thw=None,
+        video_grid_thw=None,
+        attention_mask=None,
+        **kwargs,
+    ):
+        spatial_merge_size = self.spatial_merge_size
+        mrope_position_deltas = []
+        position_ids = paddle.zeros(
+            [3, input_ids.shape[0], input_ids.shape[1]],
+            dtype=input_ids.dtype,
+        )
+
+        grid_iters = {
+            1: iter(image_grid_thw) if image_grid_thw is not None else None,
+            2: iter(video_grid_thw) if video_grid_thw is not None else None,
+        }
+
+        for batch_idx in range(input_ids.shape[0]):
+            current_input_ids = input_ids[batch_idx]
+            input_token_type = mm_token_type_ids[batch_idx]
+
+            if attention_mask is not None:
+                mask = attention_mask[batch_idx].astype("bool")
+                current_input_ids = current_input_ids[mask]
+                input_token_type = input_token_type[mask]
+
+            input_type_group = []
+            for key, group in itertools.groupby(
+                enumerate(input_token_type.tolist()), lambda x: x[1]
+            ):
+                group = list(group)
+                input_type_group.append((key, group[0][0], group[-1][0] + 1))
+
+            current_pos = 0
+            llm_pos_ids_list = []
+            for modality_type, start_idx, end_idx in input_type_group:
+                if modality_type == 0:
+                    text_len = end_idx - start_idx
+                    llm_pos_ids_list.append(
+                        paddle.arange(text_len).reshape([1, -1]).expand([3, -1])
+                        + current_pos
+                    )
+                    current_pos += text_len
+                else:
+                    grid_thw = next(grid_iters[modality_type])
+                    vision_position_ids = self.get_vision_position_ids(
+                        current_pos,
+                        grid_thw,
+                        spatial_merge_size,
+                    )
+                    llm_pos_ids_list.append(vision_position_ids)
+                    t_val = (
+                        int(grid_thw[0].item())
+                        if isinstance(grid_thw, Tensor)
+                        else int(grid_thw[0])
+                    )
+                    h_val = (
+                        int(grid_thw[1].item())
+                        if isinstance(grid_thw, Tensor)
+                        else int(grid_thw[1])
+                    )
+                    w_val = (
+                        int(grid_thw[2].item())
+                        if isinstance(grid_thw, Tensor)
+                        else int(grid_thw[2])
+                    )
+                    current_pos += max(h_val, w_val) // spatial_merge_size
+
+            llm_positions = paddle.concat(llm_pos_ids_list, axis=1).reshape(
+                [3, -1]
+            )
+
+            if attention_mask is not None:
+                mask = attention_mask[batch_idx].astype("bool")
+                position_ids[:, batch_idx, mask] = llm_positions
+            else:
+                position_ids[:, batch_idx] = llm_positions
+
+            mrope_position_deltas.append(
+                int(llm_positions.max().item()) + 1 - len(current_input_ids)
+            )
+
+        mrope_position_deltas = paddle.to_tensor(
+            mrope_position_deltas, dtype="int64"
+        ).unsqueeze(1)
+
+        return position_ids, mrope_position_deltas
+
+    def compute_3d_position_ids(
+        self,
+        input_ids=None,
+        inputs_embeds=None,
+        image_grid_thw=None,
+        video_grid_thw=None,
+        attention_mask=None,
+        past_key_values=None,
+        mm_token_type_ids=None,
+    ):
+        past_key_values_length = (
+            0
+            if past_key_values is None
+            else past_key_values.get_seq_length()
+            if hasattr(past_key_values, "get_seq_length")
+            else 0
+        )
+        can_compute_mrope = (
+            input_ids is not None
+            and mm_token_type_ids is not None
+            and (image_grid_thw is not None or video_grid_thw is not None)
+        )
+
+        if can_compute_mrope and (
+            self.rope_deltas is None or past_key_values_length == 0
+        ):
+            position_ids, rope_deltas = self.get_rope_index(
+                input_ids,
+                mm_token_type_ids=mm_token_type_ids,
+                image_grid_thw=image_grid_thw,
+                video_grid_thw=video_grid_thw,
+                attention_mask=attention_mask,
+            )
+            self.rope_deltas = rope_deltas
+            return position_ids
+
+        if self.rope_deltas is not None and inputs_embeds is not None:
+            batch_size, seq_length, _ = inputs_embeds.shape
+            if attention_mask is not None:
+                position_ids = attention_mask.astype("int64").cumsum(-1) - 1
+                position_ids = paddle.where(
+                    attention_mask == 0,
+                    paddle.zeros_like(position_ids),
+                    position_ids,
+                )
+                position_ids = position_ids.reshape([1, batch_size, -1]).tile(
+                    [3, 1, 1]
+                )
+            else:
+                position_ids = (
+                    paddle.arange(
+                        past_key_values_length,
+                        past_key_values_length + seq_length,
+                    )
+                    .reshape([1, 1, -1])
+                    .expand([3, batch_size, -1])
+                )
+
+            delta = self.rope_deltas
+            if delta.shape[0] != batch_size:
+                delta = delta.tile([batch_size // delta.shape[0], 1])
+            position_ids = position_ids + delta.unsqueeze(0)
+            return position_ids
+
+        return None
+
+    def forward(self, dict_args):
+        input_ids = dict_args.get("input_ids", None)
+        inputs_embeds = dict_args.get("inputs_embeds", None)
+        pixel_values = dict_args.get("pixel_values", None)
+        pixel_values_videos = dict_args.get("pixel_values_videos", None)
+        image_grid_thw = dict_args.get("image_grid_thw", None)
+        video_grid_thw = dict_args.get("video_grid_thw", None)
+        attention_mask = dict_args.get("attention_mask", None)
+        position_ids = dict_args.get("position_ids", None)
+        mm_token_type_ids = dict_args.get("mm_token_type_ids", None)
+        past_key_values = dict_args.get("past_key_values", None)
+
+        if (
+            inputs_embeds is None
+            and input_ids is not None
+            and self.language_model is not None
+        ):
+            inputs_embeds = self.language_embedding.embedding.embed_tokens(
+                input_ids
+            )
+
+        if pixel_values is not None and self.visual is not None:
+            image_features = self.get_image_features(
+                pixel_values, image_grid_thw
+            )
+            image_features = image_features.astype(inputs_embeds.dtype)
+            image_mask, _ = self.get_placeholder_mask(
+                input_ids,
+                inputs_embeds,
+                image_features=image_features,
+            )
+            inputs_embeds = inputs_embeds.masked_scatter(
+                image_mask, image_features
+            )
+
+        if pixel_values_videos is not None and self.visual is not None:
+            video_features = self.get_video_features(
+                pixel_values_videos, video_grid_thw
+            )
+            video_features = video_features.astype(inputs_embeds.dtype)
+            _, video_mask = self.get_placeholder_mask(
+                input_ids,
+                inputs_embeds,
+                video_features=video_features,
+            )
+            inputs_embeds = inputs_embeds.masked_scatter(
+                video_mask, video_features
+            )
+
+        if position_ids is None:
+            position_ids = self.compute_3d_position_ids(
+                input_ids=input_ids,
+                inputs_embeds=inputs_embeds,
+                image_grid_thw=image_grid_thw,
+                video_grid_thw=video_grid_thw,
+                attention_mask=attention_mask,
+                past_key_values=past_key_values,
+                mm_token_type_ids=mm_token_type_ids,
+            )
+
+        if self.config.sequence_parallel:
+            inputs_embeds = inputs_embeds.transpose([1, 0, 2]).contiguous()
+            inputs_embeds = scatter_to_sequence_parallel_region(
+                inputs_embeds, group=self.tp_group
+            )
+
+        dict_args["position_ids"] = position_ids
+        dict_args["input_ids"] = None
+        dict_args["decoder_input"] = inputs_embeds
+
+        lm_dict_args = self.language_embedding(
+            dict_args, decoder_input=inputs_embeds
+        )
+
+        for layer in self.language_backbone:
+            lm_dict_args = layer(lm_dict_args)
+
+        if self.language_lm_head is not None:
+            logits = self.language_lm_head(lm_dict_args)
+            return logits
+
+        return lm_dict_args
+
+
+# ======================================================================
+# Test dimensions
+# ======================================================================
 
 # ---- Test dimensions (small for fast unit testing) ----
 HIDDEN_SIZE = 64
@@ -40,12 +584,6 @@ IN_CHANNELS = 3
 NUM_POSITION_EMBEDDINGS = 256  # 16 * 16
 
 # Test image: 1 image, 2 temporal frames, 64x64 spatial
-#   pixel_values shape: [1, 3, 2, 64, 64]
-#   Conv3D kernel=[2,16,16] stride=[2,16,16]
-#     -> output: [1, HIDDEN_SIZE, 1, 4, 4]
-#   grid_thw = [[1, 4, 4]]
-#   seq_len = 1 * 4 * 4 = 16
-#   After spatial merge (2x2): 16 / 4 = 4 merged tokens
 IMAGE_H = 64
 IMAGE_W = 64
 GRID_T = 1
@@ -53,205 +591,6 @@ GRID_H = IMAGE_H // PATCH_SIZE  # 4
 GRID_W = IMAGE_W // PATCH_SIZE  # 4
 SEQ_LEN = GRID_T * GRID_H * GRID_W  # 16
 MERGED_TOKENS = SEQ_LEN // (SPATIAL_MERGE_SIZE**2)  # 4
-
-
-# class TestQwen3_5VisionModel(unittest.TestCase):
-#     def setUp(self):
-#         seed = 42
-#         random.seed(seed)
-#         np.random.seed(seed)
-#         paddle.seed(seed)
-
-#         strategy = fleet.DistributedStrategy()
-#         strategy.hybrid_configs = {
-#             "dp_degree": 1,
-#             "mp_degree": 1,
-#             "pp_degree": 1,
-#             "sharding_degree": 1,
-#             "sep_degree": 1,
-#             "cp_degree": 1,
-#             "ep_degree": 1,
-#             "moe_sharding_degree": 1,
-#             "order": [
-#                 "sharding",
-#                 "moe_sharding",
-#                 "pp",
-#                 "sep",
-#                 "cp",
-#                 "dp",
-#                 "ep",
-#                 "mp",
-#             ],
-#         }
-#         self.strategy = strategy
-
-#         if not ps.have_global_memory_buffer():
-#             fleet.init(is_collective=True, strategy=strategy)
-#             hcg = fleet.get_hybrid_communicate_group()
-#             ps.initialize_model_parallel(hcg)
-
-#         # Step 1: Create Qwen3.5 vision encoder
-#         config = Qwen3_5VisionProvider(
-#             num_hidden_layers=NUM_LAYERS,
-#             hidden_size=HIDDEN_SIZE,
-#             num_attention_heads=NUM_HEADS,
-#             head_dim=HEAD_DIM,
-#             out_hidden_size=OUT_HIDDEN_SIZE,
-#             intermediate_size=INTERMEDIATE_SIZE,
-#             patch_size=PATCH_SIZE,
-#             spatial_merge_size=SPATIAL_MERGE_SIZE,
-#             temporal_patch_size=TEMPORAL_PATCH_SIZE,
-#             in_channels=IN_CHANNELS,
-#             num_position_embeddings=NUM_POSITION_EMBEDDINGS,
-#             hidden_dropout_prob=0.0,
-#             attention_dropout=0.0,
-#             normalization="LayerNorm",
-#             use_qk_norm=False,
-#             gated_linear_unit=False,
-#             apply_rope_fusion=False,
-#         )
-#         self.config = config
-#         self.vision_model = config.provide()
-
-#     def test_forward(self):
-#         """Test full forward and backward of Qwen3.5 vision encoder.
-
-#         Complete computation flow starting from raw pixel input:
-#           pixel_values -> Conv3D patch embedding -> Transformer layers -> PatchMerger -> output
-#         Then backward through the entire graph, checking gradient shapes.
-#         """
-#         # ---- Step 2: create NoPipelineParallel ----
-#         vision = NoPipelineParallel(self.vision_model, self.strategy)
-
-#         # ---- Verify model structure ----
-#         layers = vision._layers.run_function
-#         print(f"\nTotal layers in run_function: {len(layers)}")
-#         for i, layer in enumerate(layers):
-#             print(f"  [{i}] {type(layer).__name__}")
-
-#         num_total = 1 + NUM_LAYERS + 1
-#         assert len(layers) == num_total, (
-#             f"Expected {num_total} layers, got {len(layers)}"
-#         )
-#         assert isinstance(layers[0], VisionEmbedding)
-#         for i in range(1, 1 + NUM_LAYERS):
-#             assert isinstance(layers[i], TransformerLayer)
-#         assert isinstance(layers[-1], Qwen3VLVisionPathMerger)
-
-#         # ---- Step 3: Forward computation (full flow from pixel input) ----
-#         # vision._layers.forward() internally does:
-#         #   x = input; for layer in run_function: x = layer(x)
-#         # We chain through run_function equivalently, calling each component's
-#         # actual sub-modules to test the complete computation graph.
-
-#         embedding = layers[0]
-#         merger = layers[-1]
-
-#         # ---- 3a. Construct raw image input ----
-#         # pixel_values: [N_chunks, C, temporal_patch_size, H, W]
-#         # For 1 image with 2 temporal frames at 64x64:
-#         pixel_values = paddle.randn(
-#             [GRID_T, IN_CHANNELS, TEMPORAL_PATCH_SIZE, IMAGE_H, IMAGE_W]
-#         )
-#         grid_thw = paddle.to_tensor(
-#             [[GRID_T, GRID_H, GRID_W]], dtype=paddle.int32
-#         )
-
-#         # ---- 3b. Patch embedding (Conv3D) ----
-#         # Conv3D: [1, 3, 2, 64, 64] -> [1, HIDDEN_SIZE, 1, 4, 4]
-#         patch_out = embedding.patch_embed(pixel_values)
-#         assert list(patch_out.shape) == [
-#             GRID_T,
-#             HIDDEN_SIZE,
-#             1,  # temporal: 2 / temporal_patch_size(2) = 1
-#             GRID_H,  # spatial: 64 / patch_size(16) = 4
-#             GRID_W,  # spatial: 64 / patch_size(16) = 4
-#         ]
-
-#         # Reshape to [1, seq_len, hidden_size]
-#         # Conv3D out: [N, C_out, D, H, W] -> flatten D/H/W -> transpose
-#         hidden_states = patch_out.flatten(2).transpose([0, 2, 1])
-#         assert list(hidden_states.shape) == [1, SEQ_LEN, HIDDEN_SIZE]
-
-#         # ---- 3c. Transformer layers ----
-#         attention_mask = paddle.ones(
-#             [1, 1, SEQ_LEN, SEQ_LEN], dtype=paddle.bool
-#         )
-#         dict_args = {
-#             "hidden_states": hidden_states,
-#             "attention_mask": attention_mask,
-#             "rotary_pos_emb": None,
-#             "rotary_pos_cos": None,
-#             "rotary_pos_sin": None,
-#             "packed_seq_params": None,
-#         }
-
-#         for i in range(1, 1 + NUM_LAYERS):
-#             dict_args = layers[i](dict_args)
-
-#         transformer_output = dict_args["hidden_states"]
-#         assert list(transformer_output.shape) == [1, SEQ_LEN, HIDDEN_SIZE], (
-#             f"Expected [1, {SEQ_LEN}, {HIDDEN_SIZE}], "
-#             f"got {list(transformer_output.shape)}"
-#         )
-
-#         # ---- 3d. Patch merger ----
-#         # Merger expects tensor [seq_len, hidden_size], not dict.
-#         # norm([16, 64]) -> reshape([-1, 256]) = [4, 256] -> MLP -> [4, 96]
-#         merger_input = transformer_output.squeeze(0)
-#         output, _ = merger(merger_input)
-#         assert list(output.shape) == [MERGED_TOKENS, OUT_HIDDEN_SIZE], (
-#             f"Expected [{MERGED_TOKENS}, {OUT_HIDDEN_SIZE}], "
-#             f"got {list(output.shape)}"
-#         )
-
-#         # ---- Step 4: Backward computation ----
-#         loss = output.sum()
-#         loss.backward()
-
-#         # Check gradients for all parameters in the forward path:
-#         #   - embedding.patch_embed (Conv3D weight/bias)
-#         #   - transformer layers (attention + MLP)
-#         #   - merger (norm + MLP)
-#         # Excluded: embedding.pos_embed, embedding.rotary_pos_emb
-#         #   (not used in this forward path)
-#         SKIP_PREFIXES = ("0.pos_embed", "0.rotary_pos_emb")
-
-#         params_with_grad = 0
-#         for name, param in vision._layers.named_parameters():
-#             if any(name.startswith(p) for p in SKIP_PREFIXES):
-#                 continue
-#             if param.grad is None:
-#                 print(f"  [NO GRAD] {name}: shape={list(param.shape)}")
-#                 continue
-
-#             params_with_grad += 1
-
-#             # Gradient shape must match parameter shape
-#             assert list(param.shape) == list(param.grad.shape), (
-#                 f"Gradient shape mismatch for {name}: "
-#                 f"param={list(param.shape)}, grad={list(param.grad.shape)}"
-#             )
-#             # Gradients must be finite
-#             assert paddle.isfinite(param.grad).all().item(), (
-#                 f"Non-finite gradients for {name}"
-#             )
-
-#             grad_norm = param.grad.detach().norm().item()
-#             print(
-#                 f"  {name}: shape={list(param.shape)}, "
-#                 f"grad_norm={grad_norm:.6f}"
-#             )
-
-#         assert params_with_grad > 0, "No parameters received gradients"
-
-
-# ======================================================================
-# Qwen3_5Model (VL composite) tests using real vision and language models
-# ======================================================================
-
-from paddlefleet.models.qwen3_5.layer_specs import get_qwen3_5_language_spec
-from paddlefleet.spec_utils import build_layer
 
 # ---- Qwen3_5Model test dimensions ----
 VL_HIDDEN_SIZE = HIDDEN_SIZE  # 64, vision hidden size
@@ -264,6 +603,11 @@ VL_TEXT_BEFORE = 5
 VL_TEXT_AFTER = 3
 VL_NUM_IMAGE_TOKENS = MERGED_TOKENS  # 4
 VL_SEQ_LEN = VL_TEXT_BEFORE + VL_NUM_IMAGE_TOKENS + VL_TEXT_AFTER  # 12
+
+
+# ======================================================================
+# Tests
+# ======================================================================
 
 
 class TestQwen3_5Model(unittest.TestCase):
@@ -374,13 +718,18 @@ class TestQwen3_5Model(unittest.TestCase):
         )
 
         # Step 3: Create Qwen3_5Model with vision and language models
-        self.model = Qwen3_5Model(
+        model = Qwen3_5Model(
             config=language_config,
             vision_model=NoPipelineParallel(vision_model, strategy),
             language_model=NoPipelineParallel(language_model, strategy),
             spatial_merge_size=SPATIAL_MERGE_SIZE,
             image_token_id=VL_IMAGE_TOKEN_ID,
             video_token_id=VL_VIDEO_TOKEN_ID,
+        )
+
+        # Convert model to bf16 (attention requires fp16/bf16 with packed_seq_params)
+        self.model = paddle.amp.decorate(
+            models=model, level="O2", dtype="bfloat16"
         )
 
     def _clear_gradients(self):
@@ -432,19 +781,13 @@ class TestQwen3_5Model(unittest.TestCase):
             "mm_token_type_ids": mm_token_type_ids,
         }
 
-        # ---- Forward ----
-        output = self.model.forward(dict_args)
+        # ---- Forward (bf16) ----
+        with paddle.amp.auto_cast(level="O2", dtype="bfloat16"):
+            output = self.model.forward(dict_args)
 
-        # assert "hidden_states" in output, "Output must contain 'hidden_states'"
-        # hidden_states = output["hidden_states"]
-        # assert list(hidden_states.shape) == [batch_size, VL_SEQ_LEN, VL_LM_HIDDEN_SIZE], (
-        #     f"Expected shape [{batch_size}, {VL_SEQ_LEN}, {VL_LM_HIDDEN_SIZE}], "
-        #     f"got {list(hidden_states.shape)}"
-        # )
-
-        # ---- Backward ----
-        loss = output.sum()
-        loss.backward()
+            # ---- Backward ----
+            loss = output.sum()
+            loss.backward()
 
         # ---- Verify gradients ----
         params_with_grad = 0
@@ -498,19 +841,13 @@ class TestQwen3_5Model(unittest.TestCase):
         input_ids = paddle.randint(0, 100, [batch_size, text_seq_len])
         dict_args = {"input_ids": input_ids}
 
-        # ---- Forward ----
-        output = self.model.forward(dict_args)
+        # ---- Forward (bf16) ----
+        with paddle.amp.auto_cast(level="O2", dtype="bfloat16"):
+            output = self.model.forward(dict_args)
 
-        # assert "hidden_states" in output
-        # hidden_states = output["hidden_states"]
-        # assert list(output.shape) == [batch_size, text_seq_len, VL_LM_HIDDEN_SIZE], (
-        #     f"Expected shape [{batch_size}, {text_seq_len}, {VL_LM_HIDDEN_SIZE}], "
-        #     f"got {list(hidden_states.shape)}"
-        # )
-
-        # ---- Backward ----
-        loss = output.sum()
-        loss.backward()
+            # ---- Backward ----
+            loss = output.sum()
+            loss.backward()
 
         # Language model params should have gradients
         lm_params_with_grad = 0
